Add: getClearTime.jsを作成した
